Implemented operators/index.js need to add types and test
