+++ src/sql_orm/select/help.js
@@ -3,3 +3,16 @@ For help query.
 
 Query to get columns and types and comments.
 */
+
+//
+
+//pull default json object
+//search postgres
+//
+
+//table/view/function comments
+//column comments
+//column types
+//grid table params
+//permissions
+//contact for help
