+++ src/sql_orm/select/search.js
@@ -1,1 +1,17 @@
 //search routes.
+
+
+
+/*
+select * 
+from ( 
+    select *, ts_vector_value
+    from schema.table
+) x
+order by x.ts_vector_value
+
+//page
+
+
+
+*/
