+++ src/sql_orm/select/where/index.js
@@ -6,9 +6,33 @@ So here's a contrived example that works for me:
 Is type conversion necessary? probably
 
 https://stackoverflow.com/questions/10720420/node-postgres-how-to-execute-where-col-in-dynamic-value-list-query
-*/
 
-/*
+
+OPERATORS: 
+
+=	Equal
+>	Greater than
+<	Less than
+>=	Greater than or equal
+<=	Less than or equal
+<> or !=	Not equal
+AND	Logical operator AND
+OR	Logical operator OR
+IN	Return true if a value matches any value in a list
+BETWEEN	Return true if a value is between a range of values
+NOT BETWEEN 
+LIKE	Return true if a value matches a pattern
+IS NULL	Return true if a value is NULL
+NOT	Negate the result of other operators
+
+a BETWEEN x AND y
+a NOT BETWEEN x AND y
+BETWEEN SYMMETRIC (does automatic swap)
+
+
+
+
+
 This module is used to help assemble the where, order_by and pagination statements for a given get request.
 CreateQueryParams is the main modules that returns the query filtering and sorting conditions string that is attached
 to the final sql query sent to the database.
@@ -153,323 +177,6 @@ function CreateQueryParamaters(req, columnObject) {
     //concat where, sort and pagination string??
 }
 
-/*
-Primary functions for creating assembly strings??
-order_statment = {variable_name: , sort_order: 'asc/desc' }
-*/
-
-
-//check if json body has all query arguments i.e. where, rules, order_by, pagination
-function CheckReqBodyArguments(req_body) {
-    CheckWhere(req_body)
-    CheckOrderBy(req_body)
-    CheckRules(req_body)
-    CheckPagniation(req_body)
-}
-
-function CheckWhere(req_body) { if (! req_body.hasOwnProperty('where')) {req_body['where'] = [{}]} }
-function CheckOrderBy(req_body) { if (! req_body.hasOwnProperty('order_by')) {req_body['order_by'] = [{}]}}
-function CheckRules(req_body) { if (! req_body.hasOwnProperty('rules')) {req_body['rules'] = [{}]}}
-function CheckPagniation(req_body) { if (! req_body.hasOwnProperty('pagination')) {
-    req_body['pagination'] = [{'offset': 0, 'limit': 1000}]}
-}
-
-
-
-//create columnObject = {columnMap, columnList, columnConcat}
-//WhereNumicialIn
-
-function WhereClause( where_statements, columnObject, replacementObject ){
-    //make aysnc for promise stuff??
-    var columnMap = columnObject['columnMap']
-    // var columnList = columnObject['columnList']
-    var columnQuickSortString = columnObject['columnQuickSortString']
-    where_list = []
-    for (where_statement of where_statements) {
-        var query_type = where_statement['query_type']
-        var data_type = where_statement['data_type']
-        var variable_name = where_statement['variable_name']
-        //determine where assembly based on query_type and data_type
-        if (['allow_update','allow_delete', 'is_assigned'].includes(variable_name)) {continue}
-        else if (query_type == 'server_query') { WhereServerQuery(where_statement, where_list) }
-        else if (query_type == 'quick_filter') { WhereQuickFilter(where_statement,  where_list, replacementObject, columnQuickSortString) }
-        else if (data_type == 'integer' || data_type == 'float' ) {
-            if (query_type == 'in') {
-                WhereNumericalIn(where_statement, columnMap, replacementObject, where_list)
-            } else {
-                WhereNumericalComparision(where_statement, columnMap, replacementObject, where_list)
-            }
-
-
-        } else if (data_type == 'string') {
-            if (query_type == 'in') {
-                // WhereIn(where_statement, columnMap, replacementObject, where_list)
-                WhereStringIn(where_statement, columnMap, replacementObject, where_list)
-            }  else if (query_type == 'not_in') {
-                WhereStringNotIn(where_statement, columnMap, replacementObject, where_list)
-            } else if (query_type == 'equals') {
-                WhereStringEqual(where_statement, columnMap, replacementObject, where_list)
-            } 
-        } else if (data_type == 'date') {
-            WhereDate(where_statement, columnMap, replacementObject, where_list)
-        } else if (data_type == 'boolean') {
-            WhereBoolean(where_statement, columnMap, replacementObject, where_list)
-        }
-
-    }
-    if (where_list.length > 0) {
-        var where_string = 'WHERE ' + where_list.join(' AND ') +'\n'
-        return where_string
-    } else { return '' }
-
-}
-
-function WhereDate(where_statement, columnMap, replacementObject, where_list) {
-    /*
-        {'before_date': YYYY-MM-DD, 'after_date': YYYY-MM-DD}
-    */
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementBeforeName = "_before_date_"+variable_name
-    var replacementAfterName = "_after_date_"+variable_name
-    var variableValue = where_statement['value']
-    // var data_type = where_statement['data_type']
-    //check date type and typecast???
-    var query_type = where_statement['query_type']
-    //need to change is null..
-    if (query_type === 'equals') {
-        if (variableValue['before_date'] !== null ) {
-            var before_date = TypeCastValues( variableValue['before_date'], 'date')
-            if (before_date === null) {return }
-            replacementObject[replacementBeforeName] = before_date
-            var boolean_string = `(${columnName} = :${replacementBeforeName})`
-            where_list.push(boolean_string)
-        }    
-    }
-
-    else if (variableValue['before_date'] !== null && variableValue['after_date'] !== null ) {
-        //date equals use same date
-        var before_date = TypeCastValues( variableValue['before_date'], 'date')
-        var after_date = TypeCastValues(variableValue['after_date'], 'date' )
-        if (before_date === null || after_date === null) {return }
-
-        replacementObject[replacementBeforeName] = before_date
-        replacementObject[replacementAfterName]  = after_date
-        if (query_type == 'between') {
-            var boolean_string = `(  :${replacementBeforeName} <= ${columnName} AND  ${columnName} <= :${replacementAfterName} )`
-            where_list.push(boolean_string)
-        } else {
-        var boolean_string = `(  ${columnName} <= :${replacementBeforeName} OR ${columnName} >= :${replacementAfterName})`
-        where_list.push(boolean_string)
-        }
-
-    } else if ( variableValue['before_date'] !== null ) {
-        var before_date = TypeCastValues( variableValue['before_date'], 'date')
-        if (before_date === null) {return }
-        replacementObject[replacementBeforeName] = before_date
-        if (query_type === 'before') {
-            var boolean_string = `(${columnName} < :${replacementBeforeName})`
-            where_list.push(boolean_string)
-        } else {
-            var boolean_string = `(${columnName} <= :${replacementBeforeName})`
-            where_list.push(boolean_string)
-        }
-
-    } else if ( variableValue['after_date'] !== null ) {
-        var after_date = TypeCastValues( variableValue['after_date'], 'date')
-        if (after_date === null) {return }
-        replacementObject[replacementAfterName] = after_date
-        if (query_type === 'after') {
-            var boolean_string = `(${columnName} > :${replacementAfterName})`
-            where_list.push(boolean_string)     
-        } else {
-            var boolean_string = `(${columnName} >= :${replacementAfterName})`
-            where_list.push(boolean_string)
-        }
-    }
-
-}
-
-//makes where in boolean statement i.e. columnName in (values)
-function WhereBoolean(where_statement, columnMap, replacementObject, where_list) {
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name
-    var variableValue = where_statement['value']
-    if (typeof variableValue !== 'boolean' ) {return }
-
-    replacementObject[variable_name] = variableValue
-    var boolean_string = `(${columnName} = ${replacementName} )`
-    where_list.push(boolean_string)
-}
-
-
-
-//makes where in boolean statement i.e. columnName in (values)
-function WhereIn(where_statement, columnMap, replacementObject, where_list) {
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name
-    var variableValue = where_statement['value']
-    var dataType = where_statement['data_type']
-    //typecast variableValue??? for integer array or string array??
-    var processedVariableValue = TypeCastValues(variableValue, dataType )
-    //if variableValue empty, null, undefined or '', skip?
-    if (IsEmptyStatement(processedVariableValue)) {return }
-    //convert to array if single value
-    if (! Array.isArray(processedVariableValue)) { processedVariableValue = [processedVariableValue]}
-    if (! Array.isArray(processedVariableValue)) { return }
-
-    replacementObject[variable_name] = processedVariableValue
-    var boolean_string = `(${columnName} in (${replacementName}) )`
-    where_list.push(boolean_string)
-}
-
-//makes where in boolean statement i.e. columnName in (values)
-function WhereStringIn(where_statement, columnMap, replacementObject, where_list) {
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name
-    var variableValue = where_statement['value']
-    var dataType = where_statement['data_type']
-    //typecast variableValue??? for integer array or string array??
-    var processedVariableValue = TypeCastValues(variableValue, dataType )
-    //if variableValue empty, null, undefined or '', skip?
-    if (IsEmptyStatement(processedVariableValue)) {return }
-    //convert to array if single value
-    if (! Array.isArray(processedVariableValue)) { processedVariableValue = [processedVariableValue]}
-    if (! Array.isArray(processedVariableValue)) { return }
-
-    replacementObject[variable_name] = processedVariableValue
-    var boolean_string = `(${columnName} ilike ANY (ARRAY [${replacementName}] ) )`
-    where_list.push(boolean_string)
-}
-
-function WhereStringNotIn(where_statement, columnMap, replacementObject, where_list) {
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name
-    var variableValue = where_statement['value']
-    var dataType = where_statement['data_type']
-    //typecast variableValue??? for integer array or string array??
-    var processedVariableValue = TypeCastValues(variableValue, dataType )
-    //if variableValue empty, null, undefined or '', skip?
-    if (IsEmptyStatement(processedVariableValue)) {return }
-    //convert to array if single value
-    if (! Array.isArray(processedVariableValue)) { processedVariableValue = [processedVariableValue]}
-    if (! Array.isArray(processedVariableValue)) { return }
-
-    replacementObject[variable_name] = processedVariableValue
-    var boolean_string = `(${columnName} not ilike ALL (ARRAY [${replacementName}] ) )`
-    where_list.push(boolean_string)
-}
-
-
-
-//Numerical In
-function WhereNumericalIn(where_statement, columnMap, replacementObject, where_list) {
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name
-    var variableValue = where_statement['value']['value_list']
-    var dataType = where_statement['data_type']
-    //typecast variableValue??? for integer array or string array??
-    var processedVariableValue = TypeCastValues(variableValue, dataType )
-    //if variableValue empty, null, undefined or '', skip?
-    if (IsEmptyStatement(processedVariableValue)) {return }
-    //convert to array if single value
-    if (! Array.isArray(processedVariableValue)) { processedVariableValue = [processedVariableValue]}
-    if (! Array.isArray(processedVariableValue)) { return }
-
-    replacementObject[variable_name] = processedVariableValue
-    var boolean_string = `(${columnName} in (${replacementName}) )`
-    where_list.push(boolean_string)
-}
-
-
-function WhereStringEqual(where_statement, columnMap, replacementObject, where_list) {
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name
-    var variableValue = where_statement['value']
-    //typecast variableValue??? for integer array or string array??
-    var processedVariableValue = TypeCastValues(variableValue, 'string' )
-    //if variableValue empty, null, undefined or '', skip?
-    if (IsEmptyStatement(processedVariableValue)) {return }
-    if (Array.isArray(processedVariableValue)) { return }
-
-    replacementObject[variable_name] = processedVariableValue
-    var boolean_string = `(${columnName} ilike ${replacementName} )`
-    where_list.push(boolean_string)
-}
-
-function WhereNumericalComparision(where_statement, columnMap, replacementObject, where_list) {
-    //variableValue: {greater: , less: , greater_equal, after_equal}
-    //greater, greater_equal, less, less_equal
-    var variable_name = where_statement['variable_name']
-    // var columnName =  columnMap[variable_name]
-    var columnName = ColumnMapReturn(columnMap, variable_name)
-    if (columnName == null) {return}
-    var replacementName = ":"+variable_name+'_1'
-    var variableValue = where_statement['value']['value_1']
-    var dataType = where_statement['data_type']
-    var query_type = where_statement['query_type']
-    //add two values to replacement object??
-
-    var processedVariableValue = TypeCastValues(variableValue, dataType )
-    if (Array.isArray(processedVariableValue)) {return }
-    if (IsEmptyStatement(processedVariableValue)) {return }
-    //need to type cast and extract values.
-    var numericalJoinList = []
-    if ( query_type == 'greater' ) {
-        numericalJoinList.push(`${columnName} > ${replacementName}`)
-    } else if ( query_type == 'less' ) {
-        numericalJoinList.push(`${columnName} < ${replacementName}`)
-    } else if ( query_type == 'greater_equal' ) {
-        numericalJoinList.push(`${columnName} >= ${replacementName}`)
-    } else if ( query_type == 'less_equal' ) {
-        numericalJoinList.push(`${columnName} <= ${replacementName}`)
-    } else if ( query_type == 'equals' ) {
-        numericalJoinList.push(`${columnName} = ${replacementName}`)
-    } 
-    
-    //include value_1 and value_2
-    else if ( query_type == 'between' || query_type == 'not_between' ) {
-        var variableValue_2 = where_statement['value']['value_2']
-        //add two values to replacement object??
-        var processedVariableValue_2 = TypeCastValues(variableValue_2, dataType )
-        if (Array.isArray(processedVariableValue_2)) {return }
-        if (IsEmptyStatement(processedVariableValue_2)) {return }
-        var replacementName_2 = ":"+variable_name+'_2'
-        replacementObject[variable_name+"_2"] = processedVariableValue_2
-        if (query_type == 'between') {
-            numericalJoinList.push(`${columnName} >= ${replacementName} AND ${columnName} <= ${replacementName_2}`)
-        } else {
-            numericalJoinList.push(`${columnName} < ${replacementName} OR ${columnName} > ${replacementName_2}`)
-        }
-    }
-
-    if (numericalJoinList.length > 0) {
-        where_list.push( '( ' + numericalJoinList.join('') + ' )' )
-    }
-    replacementObject[variable_name +"_1"] = processedVariableValue
-
-}
-
 function IsEmptyStatement(processedValue) {
     /*
     Checks if a value is null or array is empty. If either true return true
@@ -489,81 +196,5 @@ function IsEmptyStatement(processedValue) {
     return false
 }
 
-function TypeCastValues(variable_values, variable_type) {
-    /*
-    This make sure the variable type is correct. Does type conversion to a single value
-    or an array of values
-    variable_value: this is the value to be modified. Can be a single value or an array of values.
-    variable_type: The required type of the value. Can be //integer, float, date, string
 
-    if value cant be converted return null. or an empty list. Skip addition to array
-    if empty list?
-    */
-    if (variable_type == 'integer') {
-        if (IsArray(variable_values)) {
-            let mixedArray = variable_values.map(el=>parseInt(el))
-            let integerArray = mixedArray.filter( (value) => !isNaN(value) )
-            return integerArray
-        } else {
-            return parseInt(variable_values)
-        }
-
-    } else if (variable_type == 'float') {
-        if (IsArray(variable_values)) {
-            let mixedArray = variable_values.map(el=>parseFloat(el))
-            let floatArray = mixedArray.filter( (value) => !isNaN(value) )
-            return floatArray
-        } else {
-            return parseFloat(variable_values)
-        }
-
-    } else if (variable_type == 'string') {
-        if (IsArray(variable_values)) {
-            let stringArray = variable_values.map(el=> String(el))
-            return stringArray
-        } else {
-            return String(variable_values)
-        }
-
-    } else if (variable_type == 'date') {
-        //only checks for single value. No array conversion
-        var date_formats = ['YYYY-MM-DD','YYYY-M-DD','YYYY-MM-D','YYYY-M-D', 'MM/DD/YYYY','M/DD/YYYY','MM/D/YYYY','M/D/YYYY']
-        var moment_date = moment(variable_values, date_formats, true)
-        if (moment_date.isValid()) {
-            return moment_date.format('YYYY-MM-DD')
-        } else {
-            return null
-        }
-    }
-}
-
-function ColumnMapReturn(columnMap, variable_name) {
-    /*
-
-    */
-    if (columnMap.hasOwnProperty(variable_name)) {
-        return columnMap[variable_name]
-    } else {return null}
-}
-
-function IsArray(variable_values) {
-    return Array.isArray(variable_values)
-}
-
-
-//
-module.exports = {
-    'CreateQueryParamaters': CreateQueryParamaters,
-    'ExtractPaginationDataFromReq': ExtractPaginationDataFromReq,
-    'OrderClause': OrderClause,
-    'PaginationClause': PaginationClause,
-    'ExtractPaginationData': ExtractPaginationData,
-    'WhereClause': WhereClause,
-    'WhereDate': WhereDate,
-    'WhereIn': WhereIn,
-    'WhereQuickFilter': WhereQuickFilter,
-    'WhereStringEqual': WhereStringEqual,
-    'WhereServerQuery': WhereServerQuery,
-    'WhereNumericalComparision': WhereNumericalComparision,
-    'IsEmptyStatement': IsEmptyStatement
-}
+// module.exports = {}
