+++ src/sql_orm/select/where/join_condition.js
@@ -1,5 +1,8 @@
 //how to join
 
+//// AND	Logical operator AND
+// OR	Logical operator OR
+
 //array of arrays
 /*
 [
